requirements.txt
numpy
pyYAML
opencv-python
imagehash
seaborn
ijson
pycocotools
torch
torchvision
tqdm
ultralytics
Pillow
tensorboard